backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Mian Maaz Ullah Khan <115078151+mmaazkhanhere@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
langchain = "0.2.16"
langchain-google-genai = "1.0.8"
python-dotenv = "1.0.1"
pinecone-client = "5.0.1"
langchain-pinecone = "0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
